--- swift-cam/Models/AppConstants.swift
@@ -1,19 +1,0 @@
-//
-//  AppConstants.swift
-//  swift-cam
-//
-//  Application-wide constants
-//
-
-import CoreGraphics
-
-enum AppConstants {
-    static let maxClassificationResults = 5
-    static let modelSwitchDelayNanoseconds: UInt64 = 500_000_000 // 0.5 seconds
-    static let animationSpringResponse: Double = 0.8
-    static let animationDampingFraction: Double = 0.8
-    static let imageMaxHeight: CGFloat = 280
-    static let imageMinHeight: CGFloat = 200
-    static let imageMaxHeightContainer: CGFloat = 300
-}
-
