--- swift-cam/Utilities/UIImage+Extensions.swift
@@ -1,70 +1,0 @@
-//
-//  UIImage+Extensions.swift
-//  swift-cam
-//
-//  UIImage utility extensions
-//
-
-import UIKit
-import CoreVideo
-
-extension UIImage {
-    /// Converts UIImage to CVPixelBuffer for CoreML input
-    func toCVPixelBuffer() -> CVPixelBuffer? {
-        let attrs = [kCVPixelBufferCGImageCompatibilityKey: kCFBooleanTrue,
-                     kCVPixelBufferCGBitmapContextCompatibilityKey: kCFBooleanTrue] as CFDictionary
-        var pixelBuffer: CVPixelBuffer?
-        let status = CVPixelBufferCreate(kCFAllocatorDefault,
-                                       Int(self.size.width),
-                                       Int(self.size.height),
-                                       kCVPixelFormatType_32ARGB,
-                                       attrs,
-                                       &pixelBuffer)
-        
-        guard status == kCVReturnSuccess, let buffer = pixelBuffer else {
-            return nil
-        }
-        
-        CVPixelBufferLockBaseAddress(buffer, CVPixelBufferLockFlags(rawValue: 0))
-        let pixelData = CVPixelBufferGetBaseAddress(buffer)
-        
-        let rgbColorSpace = CGColorSpaceCreateDeviceRGB()
-        guard let context = CGContext(data: pixelData,
-                                    width: Int(self.size.width),
-                                    height: Int(self.size.height),
-                                    bitsPerComponent: 8,
-                                    bytesPerRow: CVPixelBufferGetBytesPerRow(buffer),
-                                    space: rgbColorSpace,
-                                    bitmapInfo: CGImageAlphaInfo.noneSkipFirst.rawValue) else {
-            CVPixelBufferUnlockBaseAddress(buffer, CVPixelBufferLockFlags(rawValue: 0))
-            return nil
-        }
-        
-        context.translateBy(x: 0, y: self.size.height)
-        context.scaleBy(x: 1.0, y: -1.0)
-        
-        UIGraphicsPushContext(context)
-        self.draw(in: CGRect(x: 0, y: 0, width: self.size.width, height: self.size.height))
-        UIGraphicsPopContext()
-        CVPixelBufferUnlockBaseAddress(buffer, CVPixelBufferLockFlags(rawValue: 0))
-        
-        return buffer
-    }
-}
-
-extension UIImage.Orientation {
-    var cgImagePropertyOrientation: CGImagePropertyOrientation {
-        switch self {
-        case .up: return .up
-        case .upMirrored: return .upMirrored
-        case .down: return .down
-        case .downMirrored: return .downMirrored
-        case .left: return .left
-        case .leftMirrored: return .leftMirrored
-        case .right: return .right
-        case .rightMirrored: return .rightMirrored
-        @unknown default: return .up
-        }
-    }
-}
-
