--- swift-cam/Views/Components/NativeCameraView.swift
@@ -1,50 +1,0 @@
-//
-//  NativeCameraView.swift
-//  swift-cam
-//
-//  Native camera picker wrapper
-//
-
-import SwiftUI
-import UIKit
-
-struct NativeCameraView: UIViewControllerRepresentable {
-    let onImageCaptured: (UIImage) -> Void
-    let onDismiss: () -> Void
-    
-    func makeUIViewController(context: Context) -> UIImagePickerController {
-        let controller = UIImagePickerController()
-        controller.sourceType = .camera
-        controller.cameraCaptureMode = .photo
-        controller.allowsEditing = false
-        controller.delegate = context.coordinator
-        return controller
-    }
-    
-    func updateUIViewController(_ uiViewController: UIImagePickerController, context: Context) {
-        // No updates needed
-    }
-    
-    func makeCoordinator() -> Coordinator {
-        Coordinator(self)
-    }
-    
-    class Coordinator: NSObject, UIImagePickerControllerDelegate, UINavigationControllerDelegate {
-        let parent: NativeCameraView
-        
-        init(_ parent: NativeCameraView) {
-            self.parent = parent
-        }
-        
-        func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey : Any]) {
-            if let image = info[.originalImage] as? UIImage {
-                parent.onImageCaptured(image)
-            }
-        }
-        
-        func imagePickerControllerDidCancel(_ picker: UIImagePickerController) {
-            parent.onDismiss()
-        }
-    }
-}
-
